refactor(app/pages): replaces all old checkbox imports;
resolves #20163

Note!!: i had to rename the Checkbox import in welcome.js, i can rename
the used component and drop the named import if needed
