+++ ui/components/app/alerts/unconnected-account-alert/unconnected-account-alert.js
@@ -18,11 +18,10 @@ import {
 import { isExtensionUrl, getURLHost } from '../../../../helpers/utils/util';
 import Popover from '../../../ui/popover';
 import Button from '../../../ui/button';
-import Checkbox from '../../../ui/check-box';
 import Tooltip from '../../../ui/tooltip';
 import ConnectedAccountsList from '../../connected-accounts-list';
 import { useI18nContext } from '../../../../hooks/useI18nContext';
-import { Icon, IconName, Text } from '../../../component-library';
+import { Icon, IconName, Text, Checkbox } from '../../../component-library';
 
 const { ERROR, LOADING } = ALERT_STATE;
