+++ ui/components/app/home-notification/home-notification.component.js
@@ -2,9 +2,8 @@ import React, { useState } from 'react';
 import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import Button from '../../ui/button';
-import Checkbox from '../../ui/check-box';
 import Tooltip from '../../ui/tooltip';
-import { Icon, IconName } from '../../component-library';
+import { Icon, IconName, Checkbox } from '../../component-library';
 import { IconColor } from '../../../helpers/constants/design-system';
 
 const HomeNotification = ({
